+++ shared/config/key.go
@@ -1,0 +1,7 @@
+package config
+
+// ProviderContextKey is the context key for the provider
+var ProviderContextKey = "provider"
+
+// OptionsContextKey is the context key for the options
+var OptionsContextKey = "options"